Refactor by converting switches to Command design pattern

+++ StmTestingSuite/Model/Command/Input/StmExternalCommandInputOption.cs
@@ -7,9 +7,9 @@ namespace StmTestingSuite.Model.Command.Input
     internal struct StmExternalCommandInputOption
     {
         public string Name { get; }
-        public ushort Value { get; }
+        public byte Value { get; }
 
-        public StmExternalCommandInputOption(string name, ushort value)
+        public StmExternalCommandInputOption(string name, byte value)
         {
             Name = name;
             Value = value;
